Conflicts:
	scilab/Makefile.in
	scilab/Version.incl
	scilab/aclocal.m4
	scilab/configure
	scilab/contrib/Makefile.in
	scilab/etc/modules.xml.in
	scilab/libs/MALLOC/Makefile.in
	scilab/libs/Makefile.in
	scilab/libs/doublylinkedlist/Makefile.in
	scilab/libs/dynamiclibrary/Makefile.in
	scilab/libs/hashtable/Makefile.in
	scilab/libs/libst/Makefile.in
	scilab/modules/Makefile.in
	scilab/modules/action_binding/Makefile.in
	scilab/modules/api_scilab/Makefile.in
	scilab/modules/arnoldi/Makefile.in
	scilab/modules/atoms/Makefile.in
	scilab/modules/boolean/Makefile.in
	scilab/modules/cacsd/Makefile.in
	scilab/modules/call_scilab/Makefile.in
	scilab/modules/compatibility_functions/Makefile.in
	scilab/modules/completion/Makefile.in
	scilab/modules/console/Makefile.in
	scilab/modules/core/Makefile.in
	scilab/modules/core/includes/callinterf.h
	scilab/modules/core/includes/version.h
	scilab/modules/core/src/c/callinterf.c
	scilab/modules/core/src/c/dynamic_gateways.c
	scilab/modules/core/src/c/dynamic_gateways.h
	scilab/modules/data_structures/Makefile.in
	scilab/modules/demo_tools/Makefile.in
	scilab/modules/development_tools/Makefile.in
	scilab/modules/development_tools/src/fake/Makefile.in
	scilab/modules/differential_equations/Makefile.in
	scilab/modules/double/Makefile.in
	scilab/modules/dynamic_link/Makefile.in
	scilab/modules/elementary_functions/Makefile.in
	scilab/modules/fftw/Makefile.in
	scilab/modules/fileio/Makefile.in
	scilab/modules/functions/Makefile.in
	scilab/modules/genetic_algorithms/Makefile.in
	scilab/modules/graphic_export/Makefile.in
	scilab/modules/graphics/Makefile.in
	scilab/modules/gui/Makefile.in
	scilab/modules/hdf5/Makefile.in
	scilab/modules/helptools/Makefile.in
	scilab/modules/history_manager/Makefile.in
	scilab/modules/integer/Makefile.in
	scilab/modules/interpolation/Makefile.in
	scilab/modules/intersci/Makefile.in
	scilab/modules/io/Makefile.in
	scilab/modules/javasci/Makefile.in
	scilab/modules/jvm/Makefile.in
	scilab/modules/linear_algebra/Makefile.in
	scilab/modules/localization/Makefile.in
	scilab/modules/m2sci/Makefile.in
	scilab/modules/maple2scilab/Makefile.in
	scilab/modules/matio/Makefile.in
	scilab/modules/mexlib/Makefile.in
	scilab/modules/optimization/Makefile.in
	scilab/modules/output_stream/Makefile.in
	scilab/modules/overloading/Makefile.in
	scilab/modules/parameters/Makefile.in
	scilab/modules/polynomials/Makefile.in
	scilab/modules/pvm/Makefile.in
	scilab/modules/randlib/Makefile.in
	scilab/modules/renderer/Makefile.in
	scilab/modules/scicos/Makefile.in
	scilab/modules/scicos_blocks/Makefile.in
	scilab/modules/shell/Makefile.in
	scilab/modules/signal_processing/Makefile.in
	scilab/modules/simulated_annealing/Makefile.in
	scilab/modules/sound/Makefile.in
	scilab/modules/sparse/Makefile.in
	scilab/modules/special_functions/Makefile.in
	scilab/modules/spreadsheet/Makefile.in
	scilab/modules/statistics/Makefile.in
	scilab/modules/string/Makefile.in
	scilab/modules/symbolic/Makefile.in
	scilab/modules/tclsci/Makefile.in
	scilab/modules/texmacs/Makefile.in
	scilab/modules/time/Makefile.in
	scilab/modules/umfpack/Makefile.in
	scilab/modules/windows_tools/Makefile.in
	scilab/modules/xcos/Makefile.in
	scilab/modules/xcos/sci_gateway/c/gw_xcos.c
	scilab/modules/xpad/Makefile.am
	scilab/modules/xpad/Makefile.in

+++ scilab/modules/xcos/src/java/org/scilab/modules/xcos/block/io/ContextUpdate.java
@@ -12,15 +12,29 @@
 
 package org.scilab.modules.xcos.block.io;
 
+import java.beans.PropertyChangeEvent;
+import java.beans.PropertyChangeListener;
 import java.io.File;
 import java.io.IOException;
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.EnumMap;
+import java.util.List;
+import java.util.Map;
+import java.util.regex.Pattern;
 
-import org.scilab.modules.graph.utils.ScilabInterpreterManagement;
-import org.scilab.modules.graph.utils.ScilabInterpreterManagement.InterpreterException;
-import org.scilab.modules.hdf5.scilabTypes.ScilabDouble;
-import org.scilab.modules.hdf5.scilabTypes.ScilabList;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.scilab.modules.action_binding.highlevel.ScilabInterpreterManagement;
+import org.scilab.modules.action_binding.highlevel.ScilabInterpreterManagement.InterpreterException;
+import org.scilab.modules.types.ScilabDouble;
+import org.scilab.modules.types.ScilabList;
+import org.scilab.modules.types.ScilabString;
+import org.scilab.modules.types.ScilabType;
 import org.scilab.modules.xcos.block.BasicBlock;
-import org.scilab.modules.xcos.io.BlockReader;
+import org.scilab.modules.xcos.block.SuperBlock;
+import org.scilab.modules.xcos.graph.SuperBlockDiagram;
+import org.scilab.modules.xcos.io.scicos.H5RWHandler;
 import org.scilab.modules.xcos.port.BasicPort;
 import org.scilab.modules.xcos.port.command.CommandPort;
 import org.scilab.modules.xcos.port.control.ControlPort;
@@ -28,69 +42,260 @@ import org.scilab.modules.xcos.port.input.ExplicitInputPort;
 import org.scilab.modules.xcos.port.input.ImplicitInputPort;
 import org.scilab.modules.xcos.port.output.ExplicitOutputPort;
 import org.scilab.modules.xcos.port.output.ImplicitOutputPort;
+import org.scilab.modules.xcos.utils.FileUtils;
+import org.scilab.modules.xcos.utils.XcosEvent;
+
+import com.mxgraph.model.mxGraphModel;
+import com.mxgraph.model.mxICell;
+import com.mxgraph.model.mxIGraphModel;
+import com.mxgraph.util.mxEventObject;
 
 /**
- * @author Clement DAVID
- *
+ * Common class for the SuperBlock I/O blocks (represent ports)
  */
 public abstract class ContextUpdate extends BasicBlock {
 
-    private static final long serialVersionUID = 6076826729067963560L;
-    
-    /**
-     * This enum represent all the subclasses of ContextUpdate .
-     *
-     * It is used to easily loop over a BasicBlock I/O blocks
-     */
-    public enum IOBlocks {
-	EventInBlock(EventInBlock.class, ControlPort.class),
-	EventOutBlock(EventOutBlock.class, CommandPort.class),
-	ExplicitInBlock(ExplicitInBlock.class, ExplicitInputPort.class),
-	ExplicitOutBlock(ExplicitOutBlock.class, ExplicitOutputPort.class),
-	ImplicitInBlock(ImplicitInBlock.class, ImplicitInputPort.class),
-	ImplicitOutBlock(ImplicitOutBlock.class, ImplicitOutputPort.class);
-	
-	private final Class< ? extends ContextUpdate> ioBlock;
-	private final Class< ? extends BasicPort> port;
-	
+	private static final Log LOG_LOCAL = LogFactory.getLog(ContextUpdate.class);
+	private static final long serialVersionUID = 6076826729067963560L;
+
 	/**
-	 * @param ioBlock input/output block
-	 * @param port the associated port class
+	 * Implement a listener which update the value and refresh the view when the
+	 * index of the port change.
 	 */
-	private IOBlocks(Class< ? extends ContextUpdate> ioBlock, Class< ? extends BasicPort> port) {
-	    this.ioBlock = ioBlock;
-	    this.port = port;
+	private static final class IndexChangeAdapter implements PropertyChangeListener, Serializable {
+		private static IndexChangeAdapter instance;
+		
+		/**
+		 * Default constructor.
+		 */
+		private IndexChangeAdapter() { }
+		
+		/**
+		 * @return the instance
+		 */
+		public static synchronized IndexChangeAdapter getInstance() {
+			if (instance == null) {
+				instance = new IndexChangeAdapter();
+			}
+			return instance;
+		}
+		
+		/**
+		 * Update the value and refresh the graph view.
+		 * 
+		 * @param evt the event
+		 * @see java.beans.PropertyChangeListener#propertyChange(java.beans.PropertyChangeEvent)
+		 */
+		@Override
+		public void propertyChange(PropertyChangeEvent evt) {
+			ScilabType data = (ScilabType) evt.getNewValue();
+			ContextUpdate ioBlock = (ContextUpdate) evt.getSource();
+			
+			if (!data.isEmpty()) {
+				int newIndex = (int) ((ScilabDouble) data).getRealPart()[0][0];
+				
+				int oldIndex;
+				if (evt.getOldValue() instanceof ScilabDouble
+						&& !((ScilabDouble) evt.getOldValue()).isEmpty()) {
+					oldIndex = (int) ((ScilabDouble) evt.getOldValue()).getRealPart()[0][0];
+				} else {
+					oldIndex = -1;
+				}
+				
+				ioBlock.setValue(newIndex);
+				
+				if (ioBlock.getParentDiagram() != null) {
+					ioBlock.getParentDiagram().fireEvent(
+							new mxEventObject(XcosEvent.IO_PORT_VALUE_UPDATED,
+									"block", ioBlock, "oldIndex", oldIndex,
+									"newIndex", newIndex));
+				}
+			}
+		}
+		
 	}
 	
 	/**
-	 * @return referenced class
+	 * Implement a listener to update the {@link ContextUpdate#isContextDependent} flag.
 	 */
-	public Class< ? extends ContextUpdate> getReferencedClass() {
-	    return ioBlock;
+	private static final class ExprsChangeAdapter implements PropertyChangeListener, Serializable {
+		private static final Pattern INTEGER_PATTERN = Pattern.compile("\\d+");
+
+		private static ExprsChangeAdapter instance;
+		
+		/**
+		 * Default constructor
+		 */
+		public ExprsChangeAdapter() { }
+		
+		/**
+		 * @return the shared instance
+		 */
+		public static ExprsChangeAdapter getInstance() {
+			if (instance == null) {
+				instance = new ExprsChangeAdapter();
+			}
+			return instance;
+		}
+		
+		/**
+		 * isContextDependant field
+		 * 
+		 * @param evt the event
+		 * @see java.beans.PropertyChangeListener#propertyChange(java.beans.PropertyChangeEvent)
+		 */
+		@Override
+		public void propertyChange(final PropertyChangeEvent evt) {
+			final ScilabType data = (ScilabType) evt.getNewValue();
+			final ContextUpdate ioBlock = (ContextUpdate) evt.getSource();
+			
+			if (!data.isEmpty()) {
+				final String newIndex = ((ScilabString) data).getData()[0][0];
+				
+				if (!INTEGER_PATTERN.matcher(newIndex).matches()) {
+					ioBlock.isContextDependent = true;
+				} else {
+					ioBlock.isContextDependent = false;
+				}
+			}
+		}
 	}
 	
 	/**
-	 * @return the port referenced class
+	 * This enum represent all the subclasses of ContextUpdate .
+	 * 
+	 * It is used to easily loop over a BasicBlock I/O blocks
 	 */
-	public Class< ? extends BasicPort> getReferencedPortClass() {
-	    return port;
+	public static enum IOBlocks {
+		/** Map a control port to an event input block */
+		EventInBlock(EventInBlock.class, ControlPort.class),
+		/** Map a command port to an event output block */
+		EventOutBlock(EventOutBlock.class, CommandPort.class),
+		/** Map an explicit input port to an explicit input block */
+		ExplicitInBlock(ExplicitInBlock.class, ExplicitInputPort.class),
+		/** Map an explicit output port to an explicit output block */
+		ExplicitOutBlock(ExplicitOutBlock.class, ExplicitOutputPort.class),
+		/** Map an implicit input port to an implicit input block */
+		ImplicitInBlock(ImplicitInBlock.class, ImplicitInputPort.class),
+		/** Map an implicit output port to an implicit output block */
+		ImplicitOutBlock(ImplicitOutBlock.class, ImplicitOutputPort.class);
+
+		private final Class< ? extends ContextUpdate> ioBlock;
+		private final Class< ? extends BasicPort> port;
+
+		/**
+		 * @param ioBlock
+		 *            input/output block
+		 * @param port
+		 *            the associated port class
+		 */
+		private IOBlocks(Class< ? extends ContextUpdate> ioBlock,
+				Class< ? extends BasicPort> port) {
+			this.ioBlock = ioBlock;
+			this.port = port;
+		}
+		
+		/**
+		 * Get all the port of the SuperBlock parent.
+		 * 
+		 * @param parent
+		 *            the parent
+		 * @return the port list mapped by port type
+		 */
+		public static Map<IOBlocks, List<mxICell>> getAllPorts(SuperBlock parent) {
+			final EnumMap<IOBlocks, List<mxICell>> ret = new EnumMap<IOBlocks, List<mxICell>>(
+					IOBlocks.class);
+
+			/* Allocation */
+			for (IOBlocks b : IOBlocks.values()) {
+				ret.put(b, new ArrayList<mxICell>());
+			}
+
+			/* Loop all over the children */
+			final int childCount = parent.getChildCount();
+
+			for (int i = 0; i < childCount; i++) {
+				final mxICell child = parent.getChildAt(i);
+
+				/* if compatible add it to the list */
+				for (IOBlocks b : IOBlocks.values()) {
+					if (child.getClass().equals(b.getReferencedPortClass())) {
+						ret.get(b).add(child);
+					}
+				}
+			}
+
+			return ret;
+		}
+
+		/**
+		 * Get all the I/O blocks of the SuperBlock parent.
+		 * 
+		 * @param parent
+		 *            the parent
+		 * @return the port list mapped by port type
+		 */
+		public static Map<IOBlocks, List<mxICell>> getAllBlocks(SuperBlock parent) {
+			final EnumMap<IOBlocks, List<mxICell>> ret = new EnumMap<IOBlocks, List<mxICell>>(
+					IOBlocks.class);
+
+			final SuperBlockDiagram graph = parent.getChild();
+			if (graph == null) {
+				parent.createChildDiagram();
+			}
+			
+			/* Allocation */
+			for (IOBlocks b : IOBlocks.values()) {
+				ret.put(b, new ArrayList<mxICell>());
+			}
+
+			/* Loop all over the children */
+			final Object defaultParent = graph.getDefaultParent();
+			final mxIGraphModel defaultModel = graph.getModel();
+			final Object[] children = mxGraphModel.getChildCells(
+					defaultModel, defaultParent, true, false);
+			
+			for (Object element : children) {
+				final mxICell child = (mxICell) element;
+
+				/* if compatible add it to the list */
+				for (IOBlocks b : IOBlocks.values()) {
+					if (child.getClass().equals(b.getReferencedClass())) {
+						ret.get(b).add(child);
+					}
+				}
+			}
+
+			return ret;
+		}
+
+		/**
+		 * @return referenced class
+		 */
+		public Class< ? extends ContextUpdate> getReferencedClass() {
+			return ioBlock;
+		}
+
+		/**
+		 * @return the port referenced class
+		 */
+		public Class< ? extends BasicPort> getReferencedPortClass() {
+			return port;
+		}
 	}
 	
-    }
-    
+	private transient boolean isContextDependent;
+	
 	/**
-	 * Constructor
+	 * Constructor.
 	 */
 	public ContextUpdate() {
 		super();
-	}
-
-	/**
-	 * @param label block labek
-	 */
-	protected ContextUpdate(String label) {
-		this();
-		setValue(label);
+		
+		getParametersPCS().addPropertyChangeListener(INTEGER_PARAMETERS,
+				IndexChangeAdapter.getInstance());
+		getParametersPCS().addPropertyChangeListener(EXPRS,
+				ExprsChangeAdapter.getInstance());
 	}
 
 	/**
@@ -104,45 +309,50 @@ public abstract class ContextUpdate extends BasicBlock {
 		setODState(new ScilabList());
 		setValue(1);
 	}
-    
-    /**
-     * @param context new context
-     */
-    public void onContextChange(String[] context) {
-	//prevent to open twice
-	if (isLocked()) {
-	    return;
-	}
-	
-	final File tempOutput;
-	final File tempInput;
-	final File tempContext;
-	try {
-	    tempInput = File.createTempFile("xcos", ".h5");
-	    tempInput.deleteOnExit();
 
-	    // Write scs_m
-	    tempOutput = exportBlockStruct();
-	    // Write context
-	    tempContext = exportContext(context);
-
-	    String cmd;
-	    
-	    cmd = "xcosBlockEval(\"" + tempOutput.getAbsolutePath() + "\"";
-	    cmd += ", \"" + tempInput.getAbsolutePath() + "\"";
-	    cmd += ", " + getInterfaceFunctionName();
-	    cmd += ", \"" + tempContext.getAbsolutePath() + "\");";
-
-		try {
-			ScilabInterpreterManagement.synchronousScilabExec(cmd);
-		} catch (InterpreterException e) {
-			e.printStackTrace();
+	/**
+	 * @param context
+	 *            new context
+	 */
+	public void onContextChange(String[] context) {
+		// prevent to open twice
+		if (isLocked()) {
+			return;
 		}
-		BasicBlock modifiedBlock = BlockReader.readBlockFromFile(tempInput.getAbsolutePath());
-		updateBlockSettings(modifiedBlock);
-	    
-	} catch (IOException e) {
-	    e.printStackTrace();
+
+		// do not evaluate context is the block is not context dependent.
+		if (!isContextDependent) {
+			return;
+		}
+		
+		LOG_LOCAL.trace("Update the I/O value from the context");
+		
+		File tempOutput;
+		File tempInput;
+		File tempContext;
+		try {
+			tempInput = FileUtils.createTempFile();
+			tempInput.deleteOnExit();
+
+			// Write scs_m
+			tempOutput = exportBlockStruct();
+			// Write context
+			tempContext = exportContext(context);
+
+			String cmd = ScilabInterpreterManagement.buildCall("xcosBlockEval",
+					tempOutput.getAbsolutePath(), tempInput.getAbsolutePath(),
+					getInterfaceFunctionName().toCharArray(), tempContext.getAbsolutePath());
+
+			try {
+				ScilabInterpreterManagement.synchronousScilabExec(cmd);
+			} catch (InterpreterException e) {
+				e.printStackTrace();
+			}
+			BasicBlock modifiedBlock = new H5RWHandler(tempInput).readBlock();
+			updateBlockSettings(modifiedBlock);
+
+		} catch (IOException e) {
+			LOG_LOCAL.error(e);
+		}
 	}
-    }
 }
